Cleaned up the directories and also broke a part the chore components to make it easier to follow what each thing is doing. Consolidated components into a Volunteer and Request Landing Pages as well
