+++ client/src/components/chore-components/RequestNotAccepted.js
@@ -9,12 +9,30 @@ const Chore = props => (
         <td className={props.chore.chore_completed ? 'completed' : ''}>{props.chore.chore_responsible}</td>
         <td className={props.chore.chore_completed ? 'completed' : ''}>{props.chore.chore_address[0].formattedAddress}</td>
         <td className={props.chore.chore_completed ? 'completed' : ''}>{props.chore.chore_phone}</td>
-        
+        <td>
+            <Link to={"/edit/"+props.chore._id}>Edit</Link>
+        </td> 
        
     </tr>
 )
 
-export default class VolunteeCompletedList extends Component {
+const RequestedChore = props => (
+    <tr>
+        <td>{props.chore.chore_description}</td>
+        <td>{props.chore.chore_responsible}</td>
+        <td>{props.chore.chore_address[0].formattedAddress}</td>
+        <td>{props.chore.chore_phone}</td>
+        <td>
+            <button type="button" class="btn btn-primary" onClick={()=>props.onAccept(props.chore)}>Accept</button>
+        </td> 
+       
+    </tr>
+)
+
+
+
+
+export default class ChoresList extends Component {
 
     constructor(props) {
         super(props);
@@ -50,12 +68,10 @@ export default class VolunteeCompletedList extends Component {
         });
     }
 
-    
-
-    completedList() {
-        const filteredListCompleted = this.getList(true, true);
-        return filteredListCompleted.map(function(currentChore, i) {
-            return <Chore chore={currentChore} key={i} />;
+    requestedList() {
+        const filteredListRequested = this.getList(false, false);
+        return filteredListRequested.map((currentChore, i) => {
+            return <RequestedChore chore={currentChore} key={i} onAccept={this.onAccept} />;
         });
     }
 
@@ -98,7 +114,7 @@ export default class VolunteeCompletedList extends Component {
     render() {
         return (
             <div>
-                <h3>Completed Chores</h3>
+                <h3>Requested Chores</h3>
                 <table className="table table-striped" style={{ marginTop: 20 }}>
                     <thead>
                         <tr>
@@ -106,16 +122,15 @@ export default class VolunteeCompletedList extends Component {
                             <th>Responsible</th>
                             <th>Address</th>
                             <th>Contact Phone</th>
+                            <th>Actions</th>
                         </tr>
                     </thead>
                     <tbody>
-                        { this.completedList() }
+                        { this.requestedList() }
                     </tbody>
                 </table>
 
-            </div>
-
-            
+            </div>            
         )
     }
 }
