+++ client/src/components/map-components/VolunteerMapContainer.js
@@ -1,18 +1,11 @@
-
 import React, { Component } from 'react'
 import {  } from 'react-google-maps'
-import ChoresMap from '../components/map-components/ChoresMap'
-import VolunteerAcceptedList from '../components/chore-components/VolunteerAcceptedList';
-import VolunteerCompletedList from '../components/chore-components/VolunteerCompletedList';
-import Card from '@material-ui/core/Card'
+import ChoresMap from './ChoresMap'
 import { filter, cloneDeep, map } from 'lodash';
 import axios from "axios";
-import CssBaseline from '@material-ui/core/CssBaseline';
 
 
-
-
-class MapView extends Component {
+class VolunteerMapContainer extends Component {
   constructor(props){
     super(props)
     this.state = {
@@ -96,31 +89,20 @@ class MapView extends Component {
 
   render() {
     return (
-      <React.Fragment>
-        <CssBaseline/>
-        <Card style={{padding: 20, margin: 20}}>  
-          <ChoresMap
-            isMarkerShown={this.state.isMarkerShown}
-            currentLocation={this.state.currentLatLng}
-            chores={this.state.chores}
-            activeMarker={this.state.activeMarker}
-            closeOtherMarkers={this.closeOtherMarkers}
-            onAccept={this.onAccept}
-          />
-          
-        </Card>
-        <Card style={{padding: 20, margin: 20}}>
-          <VolunteerAcceptedList/>
-        </Card> 
-        <Card style={{padding: 20, margin: 20}}>
-          <VolunteerCompletedList/>
-        </Card> 
-      </React.Fragment>    
+       
+    <ChoresMap
+      isMarkerShown={this.state.isMarkerShown}
+      currentLocation={this.state.currentLatLng}
+      chores={this.state.chores}
+      activeMarker={this.state.activeMarker}
+      closeOtherMarkers={this.closeOtherMarkers}
+      onAccept={this.onAccept}
+    />     
     )
   }
 }
 
-export default MapView;
+export default VolunteerMapContainer;
